\name{sara4r}
\alias{sara4r}
\title{An R-GUI for Spatial Analysis of Surface Runoff using the NRCS-CN Method}
\description{a raster-based GIS tool that outputs runoff estimates from Land use/land cover and hydrologic soil group maps.
}
\usage{
sara4r()
}
\details{
\tabular{ll}{
Package: \tab sara4r\cr
Type: \tab Package\cr
Version: \tab 0.0.7\cr
Date: \tab 2020-01-14\cr
Depends: \tab R(>= 3.4.0),RGtk2\cr
Imports: \tab raster,sp,rgdal\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
}
}
\references{
[CN-Idris, Hernández-Guzmán et al., 2011 - CN-Idris: An Idrisi tool for generating curve number maps and estimating direct runoff. Environmental Modelling & Software, 26(12), 1764-1766](https://doi.org/10.1016/j.envsoft.2011.07.006)

[SARA, Hernández-Guzmán and Ruiz-Luna, 2013. SARA – An enhanced curve number-based tool
for estimating direct runoff. Journal of Hydroinformatics, 15(3), 881-887](https://doi.org/10.2166/hydro.2013.145)
}
\author{Rafael Hernandez Guzman, Arturo Ruiz Luna\cr
Maintainer: Rafael Hernandez Guzman<rhernandez.g@gmail.com>
}
\note{
http://hydro-geomatic-lab.com/sara4r.html
}
\examples{
sara4r()
}

