\name{sara4r}
\alias{sara4r}
\title{An R-GUI for Spatial Analysis of Surface Runoff using the NRCS-CN Method}
\description{a raster-based GIS tool that outputs runoff estimates from Land use/land cover and hydrologic soil group maps.
}
\usage{
sara4r()
}
\details{
\tabular{ll}{
Package: \tab sara4r\cr
Type: \tab Package\cr
Version: \tab 0.0.9\cr
Date: \tab 2022-01-31\cr
Depends: \tab R(>= 4.1.2), tcltk, tcltk2\cr
Imports: \tab raster,sp,rgdal\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
}
}
\references{
[CN-Idris, Hernández-Guzmán et al., 2011 - CN-Idris: An Idrisi tool for generating curve number maps and estimating direct runoff. Environmental Modelling & Software, 26(12), 1764-1766](https://doi.org/10.1016/j.envsoft.2011.07.006)

[SARA, Hernández-Guzmán and Ruiz-Luna, 2013. SARA – An enhanced curve number-based tool
for estimating direct runoff. Journal of Hydroinformatics, 15(3), 881-887](https://doi.org/10.2166/hydro.2013.145)

[SARA4R, Hernández-Guzmán et al., 2021. Sara4r – an R graphical user interface (GUI) to estimate watershed surface runoff applying the NRCS – curve number method. Journal of Hydroinformatics, 23(1), 76-87](https://doi.org/10.2166/hydro.2020.087)
}
\author{Rafael Hernandez Guzman, Arturo Ruiz Luna\cr
Maintainer: Rafael Hernandez Guzman<rhernandez.g@gmail.com>
}
\note{
http://hydro-geomatic-lab.com/sara4r.html
}
\examples{
sara4r()
}

